visa center models separated to a directory, passport model added without migrations and relations
