+++ visa_center/requests.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, Field
 
-from visa_center.models import CountryEnum
+from visa_center.models.visa_center_model import CountryEnum
 
 
 class AddVisaAccountRequest(BaseModel):
